// client/src/app/models/usuario.model.ts
export class UsuarioModel{
    
    rol:String;
    primerNombre: String;
    segundoNombre: String;
    primerApellido: String;
    segundoApellido: String;
    email: String;
    contrasena: String;
    fechaNacimiento: String;
    direccion: String;
    ciudad: String;
    departamento: String;
    telefono: String;
    isLogged: boolean= false;
    listaInmuebles:String [];

}